Add Dockerfiles/docker-compose and requirements for API, front-end, and authenticator; implement new templates for user interactions
