+++ authenticator/requirements.txt
@@ -1,0 +1,4 @@
+fastapi
+uvicorn
+pyJWT
+python-multipart
