+++ docker-compose.yml
@@ -1,0 +1,32 @@
+version: '3.8'
+
+services:
+  front:
+    build: ./front
+    ports:
+      - "5000:5000"
+    depends_on:
+      - api
+      - authenticator
+    networks:
+      - reseau1
+
+  api:
+    build: ./api
+    ports:
+      - "5001:5001"
+    depends_on:
+      - authenticator
+    networks:
+      - reseau1
+
+  authenticator:
+    build: ./authenticator
+    ports:
+      - "5002:5002"
+    networks:
+      - reseau1
+
+networks:
+  reseau1:
+    driver: bridge
