+++ front/templates/index.j2
@@ -1,0 +1,85 @@
+{% import "utils.j2" as utils %}
+
+<!DOCTYPE html>
+<html lang="en">
+    {{ utils.head("Accueil") }}
+    <body>
+        {{ utils.jumbotron_top("Accueil") }}
+        {{ utils.navbar([{"href":"/", "name":"Accueil"}, {"href":"/test", "name":"test"}]) }}
+
+
+        <div class="container" style="margin-top:30px">
+            <h3 style="text-align: center;">Bienvenue sur le site de la librairie !</h3><br><br>
+            <div class="row">
+                
+                <div class="col-sm-3">
+                    <p>Vous rechercher : </p>
+
+                    <ul>
+                        <li>Un livre </li>
+                        <button class="btn btn-primary" onclick="location.href='/rechercher?type=livres'">Rechercher</button>
+                    </ul>
+
+                    <ul>
+                        <li>Un utilisateur </li>
+                        <button class="btn btn-primary" onclick="location.href='/rechercher?type=utilisateurs'">Rechercher</button>
+                    </ul>
+
+                    <ul>
+                        <li>Un auteur </li>
+                        <button class="btn btn-primary" onclick="location.href='/rechercher?type=auteurs'">Rechercher</button>
+                    </ul>
+
+                </div>
+
+
+                <div class="col-sm-6">
+                    <h4>Formulaire : trouver si un livre est emprunté</h1>
+                    <form action="{{ url_for('resultats') }}" method="POST" onsubmit="return validateForm()">
+                        <div>
+                        <label for="utilisateur">Utilisateur:</label>
+                        <input class="form-control" type="text" id="utilisateur" name="utilisateur">
+                        <button class="form-control" type="submit" name="action" value="submit_utilisateur">Submit Utilisateur</button>
+                        </div>
+                        <div>
+                        <label for="livre">Livre:</label>
+                        <input class="form-control" type="text" id="livre" name="livre">
+                        <button class="form-control" type="submit" name="action" value="submit_livre">Submit Livre</button>
+                        </div>
+                    </form>
+
+                    <script>
+                        function validateForm() {
+                        var utilisateur = document.getElementById("utilisateur").value;
+                        var livre = document.getElementById("livre").value;
+                        if (utilisateur === "" && livre === "") {
+                            alert("Veuillez remplir au moins un des champs.");
+                            return false;
+                        }
+                        return true;
+                        }
+                    </script>
+                </div>
+
+                <div class="col-sm-3">
+                    <p>Vous souhaitez ajouter : </p>
+
+                    <ul>
+                        <li>Un livre </li>
+                        <button class="btn btn-primary" onclick="location.href='/ajouter?type=livres'">ajouter</button>
+                    </ul>
+
+                    <ul>
+                        <li>Un utilisateur </li>
+                        <button class="btn btn-primary" onclick="location.href='/ajouter?type=utilisateurs'">ajouter</button>
+                    </ul>
+
+                    <ul>
+                        <li>Un auteur </li>
+                        <button class="btn btn-primary" onclick="location.href='/ajouter?type=auteurs'">ajouter</button>
+                    </ul>
+            </div>
+        </div>
+
+    </body>
+</html>
