+++ front/templates/lettres.j2
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="fr">
+    <head>
+        <meta charset="UTF-8">
+        <title>Liste de mots</title>
+    </head>
+    <body>
+        <h1>{{ label }} : {{ lettre_fournie }}</h1>
+        <textarea style="width: 300px; height: 200px; overflow-y: scroll;" readonly>
+            {% for each in mots %}
+                {{ each }}
+            {% endfor %}
+        </textarea>
+        <br>
+        <a href={{ url_for('index') }}>Retour</a>
+    </body>
+</html>
