+++ front/templates/resultats.j2
@@ -1,0 +1,27 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="UTF-8">
+    <title>Résultats des Emprunts</title>
+    <style>
+        body { font-family: Arial, sans-serif; margin: 20px; }
+        .error { color: red; }
+        .livre { margin-bottom: 10px; }
+    </style>
+</head>
+<body>
+    <h1>Emprunts pour l'utilisateur : {{ utilisateur }}</h1>
+
+    {% if error %}
+        <p class="error">Erreur : {{ error }}</p>
+    {% elif emprunt %}
+        <h2>Liste des Livres Empruntés :</h2>
+        <ul>
+            {{ emprunt }}
+        </ul>
+    {% else %}
+        <p>Aucun emprunt trouvé pour cet utilisateur.</p>
+    {% endif %}
+    <a href="/accueil">Retour au formulaire</a>
+</body>
+</html>
